【ID1000038】 商家板块

+++ src/main/java/com/omate/liuqu/repository/EventRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface EventRepository extends CrudRepository<Event, Integer> {
+public interface EventRepository extends CrudRepository<Event, Long> {
     List<Event> findAll();
 }
